Backend Milestone

+++ taskfoo-backend/src/main/java/com/taskfoo/taskfoo_backend/service/TaskService.java
@@ -3,6 +3,7 @@ package com.taskfoo.taskfoo_backend.service;
 import com.taskfoo.taskfoo_backend.model.Task;
 import com.taskfoo.taskfoo_backend.repository.TaskRepository;
 import org.springframework.stereotype.Service;
+import jakarta.persistence.EntityNotFoundException;
 
 import java.util.List;
 
@@ -10,6 +11,10 @@ import java.util.List;
 public class TaskService {
 
     private final TaskRepository taskRepository;
+
+    public List<Task> searchTasks(String keyword) {
+        return taskRepository.findByTitleContainingIgnoreCase(keyword);
+    }
 
     public TaskService(TaskRepository taskRepository) {
         this.taskRepository = taskRepository;
@@ -20,7 +25,8 @@ public class TaskService {
     }
 
     public Task getTaskById(Long id) {
-        return taskRepository.findById(id).orElse(null);
+        return taskRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Task not found"));
     }
 
     public Task createTask(Task task) {
@@ -28,22 +34,24 @@ public class TaskService {
     }
 
     public Task updateTask(Long id, Task updatedTask) {
-        Task existingTask = taskRepository.findById(id).orElse(null);
-        if (existingTask != null) {
-            existingTask.setTitle(updatedTask.getTitle());
-            existingTask.setDescription(updatedTask.getDescription());
-            existingTask.setStatus(updatedTask.getStatus());
-            existingTask.setPriority(updatedTask.getPriority());
-            existingTask.setEpic(updatedTask.getEpic());
-            existingTask.setStartDate(updatedTask.getStartDate());
-            existingTask.setDueDate(updatedTask.getDueDate());
-            existingTask.setCreatedAt(updatedTask.getCreatedAt());
-            return taskRepository.save(existingTask);
-        }
-        return null;
+        Task existingTask = taskRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Task not found"));
+
+        existingTask.setTitle(updatedTask.getTitle());
+        existingTask.setDescription(updatedTask.getDescription());
+        existingTask.setStatus(updatedTask.getStatus());
+        existingTask.setPriority(updatedTask.getPriority());
+        existingTask.setEpic(updatedTask.getEpic());
+        existingTask.setStartDate(updatedTask.getStartDate());
+        existingTask.setDueDate(updatedTask.getDueDate());
+
+        return taskRepository.save(existingTask);
     }
 
     public void deleteTask(Long id) {
+        if (!taskRepository.existsById(id)) {
+            throw new EntityNotFoundException("Task not found");
+        }
         taskRepository.deleteById(id);
     }
 }
